--- Source/JoltExample/Public/JoltPhysicsTest.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "CoreMinimal.h"
-
-//#include "JoltPhysicsTest.generated.h"
-
-class JoltPhysicsTest{
-
-public:
-	JoltPhysicsTest();
-	~JoltPhysicsTest();
-};
